refactor: universal email customisation and handling
