+++ includes/class-newspack.php
@@ -73,6 +73,7 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/class-logger.php';
 
 		include_once NEWSPACK_ABSPATH . 'includes/util.php';
+		include_once NEWSPACK_ABSPATH . 'includes/emails/class-emails.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-plugin-manager.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-theme-manager.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-admin-plugins-screen.php';
