+++ includes/wizards/class-reader-revenue-wizard.php
@@ -534,7 +534,7 @@ class Reader_Revenue_Wizard extends Wizard {
 			'newspack-reader-revenue-wizard',
 			'newspack_reader_revenue',
 			[
-				'emails'                  => Reader_Revenue_Emails::get_emails(),
+				'emails'                  => Emails::get_emails( [ Reader_Revenue_Emails::EMAIL_TYPES['RECEIPT'] ] ),
 				'salesforce_redirect_url' => Salesforce::get_redirect_url(),
 			]
 		);
