+++ tests/unit-tests/emails.php
@@ -6,17 +6,29 @@
  */
 
 use Newspack\Plugin_Manager;
-use Newspack\Reader_Revenue_Emails;
+use Newspack\Emails;
 
 /**
  * Tests Reader Revenue Emails.
  */
-class Newspack_Test_Reader_Revenue_Emails extends WP_UnitTestCase {
+class Newspack_Test_Emails extends WP_UnitTestCase {
 	/**
 	 * Setup.
 	 */
 	public function setUp() {
 		reset_phpmailer_instance();
+		add_filter(
+			'newspack_email_configs',
+			function ( $types ) {
+				$types['test-email-config'] = [
+					'name'        => 'test-email-config',
+					'label'       => __( 'Test config', 'newspack' ),
+					'description' => __( 'Email sent to test things.', 'newspack' ),
+					'template'    => dirname( NEWSPACK_PLUGIN_FILE ) . '/includes/templates/reader-revenue-emails/receipt.php',
+				];
+				return $types;
+			}
+		);
 	}
 
 	/**
@@ -31,76 +43,76 @@ class Newspack_Test_Reader_Revenue_Emails extends WP_UnitTestCase {
 	 *
 	 * @param string $type Email type.
 	 */
-	private static function get_email( $type ) {
-		$email = Reader_Revenue_Emails::get_emails()[ $type ];
+	private static function get_test_email( $type ) {
+		$email = Emails::get_emails()[ $type ];
 		// Mitigate a weird issue with PHPUnit – when running the whole test suite (as opposed to a
 		// single run with `--filter` flag), the default values from `register_meta` calls are ignored.
 		update_post_meta( $email['post_id'], 'from_name', get_bloginfo( 'name' ) );
 		update_post_meta( $email['post_id'], 'from_email', get_bloginfo( 'admin_email' ) );
 		// Return the updated version.
-		return Reader_Revenue_Emails::get_emails()['receipt'];
+		return Emails::get_emails()[ $type ];
 	}
 
 	/**
 	 * Email setup & defaults generation.
 	 */
-	public function test_reader_revenue_emails_setup() {
+	public function test_emails_setup() {
 		self::assertEquals(
-			Reader_Revenue_Emails::get_emails(),
+			Emails::get_emails( [ 'test-email-config' ] ),
 			[],
 			'Emails are empty until configured.'
 		);
 		self::assertFalse(
-			Reader_Revenue_Emails::can_send_email( 'receipt' ),
-			'Receipt email cannot be sent.'
+			Emails::can_send_email( 'test-email-config' ),
+			'Test email cannot be sent.'
 		);
 		self::assertFalse(
-			Reader_Revenue_Emails::supports_emails(),
+			Emails::supports_emails(),
 			'Emails are not configured until the Newspack Newsletters plugin is active.'
 		);
-		$send_result = Reader_Revenue_Emails::send_email(
-			Reader_Revenue_Emails::EMAIL_TYPE_RECEIPT,
+		$send_result = Emails::send_email(
+			'test-email-config',
 			'someone@example.com'
 		);
 		self::assertFalse( $send_result, 'Email cannot be sent until the instance is configured.' );
 
 		Plugin_Manager::activate( 'newspack-newsletters' );
 		self::assertTrue(
-			Reader_Revenue_Emails::supports_emails(),
+			Emails::supports_emails(),
 			'Emails are configured after Newspack Newsletters plugin is active.'
 		);
 
 		self::assertTrue(
-			Reader_Revenue_Emails::can_send_email( 'receipt' ),
-			'Receipt email can now be sent.'
+			Emails::can_send_email( 'test-email-config' ),
+			'Test email can now be sent.'
 		);
 
-		$emails        = Reader_Revenue_Emails::get_emails();
-		$receipt_email = $emails['receipt'];
+		$emails     = Emails::get_emails( [ 'test-email-config' ] );
+		$test_email = $emails['test-email-config'];
 		self::assertEquals(
-			'Receipt',
-			$receipt_email['label'],
-			'Receipt email has the expected label'
+			'Test config',
+			$test_email['label'],
+			'Test email has the expected label'
 		);
 		self::assertEquals(
 			'Thank you!',
-			$receipt_email['subject'],
-			'Receipt email has the expected subject'
+			$test_email['subject'],
+			'Test email has the expected subject'
 		);
 		self::assertContains(
 			'<!doctype html>',
-			$receipt_email['html_payload'],
-			'Receipt email has the HTML payload'
+			$test_email['html_payload'],
+			'Test email has the HTML payload'
 		);
 	}
 
 	/**
-	 * Email sending.
+	 * Email sending, with a template.
 	 */
-	public function test_reader_revenue_emails_send() {
+	public function test_emails_send_with_template() {
 		Plugin_Manager::activate( 'newspack-newsletters' );
 
-		$receipt_email = self::get_email( 'receipt' );
+		$test_email = self::get_test_email( 'test-email-config' );
 
 		$recipient    = 'tester@tests.com';
 		$amount       = '$42';
@@ -110,8 +122,8 @@ class Newspack_Test_Reader_Revenue_Emails extends WP_UnitTestCase {
 				'value'    => $amount,
 			],
 		];
-		$send_result  = Reader_Revenue_Emails::send_email(
-			Reader_Revenue_Emails::EMAIL_TYPE_RECEIPT,
+		$send_result  = Emails::send_email(
+			'test-email-config',
 			$recipient,
 			$placeholders
 		);
@@ -126,7 +138,7 @@ class Newspack_Test_Reader_Revenue_Emails extends WP_UnitTestCase {
 			'Sent email has the expected recipient'
 		);
 		self::assertEquals(
-			$receipt_email['subject'],
+			$test_email['subject'],
 			$mailer->get_sent()->subject,
 			'Sent email has the expected subject'
 		);
@@ -145,39 +157,39 @@ class Newspack_Test_Reader_Revenue_Emails extends WP_UnitTestCase {
 	/**
 	 * Sending by email id.
 	 */
-	public function test_reader_revenue_emails_send_by_id() {
+	public function test_emails_send_by_id() {
 		Plugin_Manager::activate( 'newspack-newsletters' );
-		$receipt_email = self::get_email( 'receipt' );
+		$test_email = self::get_test_email( 'test-email-config' );
 
-		$send_result = Reader_Revenue_Emails::send_email(
-			$receipt_email['post_id'],
+		$send_result = Emails::send_email(
+			$test_email['post_id'],
 			'someone@example.com'
 		);
 		self::assertTrue( $send_result, 'Email has been sent.' );
 
-		$send_result = Reader_Revenue_Emails::send_email(
+		$send_result = Emails::send_email(
 			9999,
 			'someone@example.com'
 		);
-		self::assertFalse( $send_result, 'Non-existend email is not sent.' );
+		self::assertFalse( $send_result, 'Non-existent email is not sent.' );
 	}
 
 	/**
 	 * Email post status handling.
 	 */
-	public function test_reader_revenue_emails_status() {
+	public function test_emails_status() {
 		Plugin_Manager::activate( 'newspack-newsletters' );
-		$receipt_email = self::get_email( 'receipt' );
+		$test_email = self::get_test_email( 'test-email-config' );
 		wp_update_post(
 			[
-				'ID'          => $receipt_email['post_id'],
+				'ID'          => $test_email['post_id'],
 				'post_status' => 'draft',
 			]
 		);
 
-		self::assertFalse( Reader_Revenue_Emails::can_send_email( 'receipt' ), 'Email can\'t be sent – it\'s not published.' );
-		$send_result = Reader_Revenue_Emails::send_email(
-			$receipt_email['post_id'],
+		self::assertFalse( Emails::can_send_email( 'test-email-config' ), 'Email can\'t be sent – it\'s not published.' );
+		$send_result = Emails::send_email(
+			$test_email['post_id'],
 			'someone@example.com'
 		);
